requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
requests==2.24.0
torch==1.11.0+cpu
torchvision==0.12.0+cpu
urllib3==1.25.10
Werkzeug==1.0.1
anyio==3.6.1
colorama==0.4.5
fastapi==0.79.0
h11==0.13.0
pydantic==1.9.1
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
typing_extensions==4.3.0
uvicorn==0.18.2
python-multipart
matplotlib
seaborn
streamlit
gpiozero
picamera
scikit-image
scikit-learn
numpy
opendatasets
Pillow